// apps/frontend/src/app/core/api/models/column-filter.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FilterOperator } from '../models/filter-operator'

/**
 * Filter applied to a single column.
 */
export interface ColumnFilter {
  /**
   * Filter comparison operator
   */
  operator: FilterOperator

  /**
   * Filter value to match against
   */
  value: string
}